Rename dataDetails as clusterDetails for use with OCP details empty state
