+++ src/routes/details/components/providerDetails/clusterDetails/clusterDetails.styles.ts
