+++ src/routes/details/components/providerDetails/clusterDetails/clusterDetails.tsx
@@ -3,10 +3,10 @@ import { Modal, ModalBody, ModalHeader, ModalVariant } from '@patternfly/react-c
 import messages from 'locales/messages';
 import React, { useState } from 'react';
 import { useIntl } from 'react-intl';
-import { OverallStatus } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/components/overallStatus';
+import { OverallStatus } from 'routes/details/components/providerDetails/clusterDetails/components/overallStatus';
 
-import { styles } from './dataDetails.styles';
-import { DataDetailsContent } from './dataDetailsContent';
+import { styles } from './clusterDetails.styles';
+import { ClusterDetailsContent } from './clusterDetailsContent';
 
 interface DataDetailsOwnProps {
   clusterId?: string;
@@ -14,7 +14,7 @@ interface DataDetailsOwnProps {
 
 type DataDetailsProps = DataDetailsOwnProps;
 
-const DataDetails: React.FC<DataDetailsProps> = ({ clusterId }: DataDetailsProps) => {
+const ClusterDetails: React.FC<DataDetailsProps> = ({ clusterId }: DataDetailsProps) => {
   const intl = useIntl();
   const [isOpen, setIsOpen] = useState(false);
 
@@ -38,11 +38,11 @@ const DataDetails: React.FC<DataDetailsProps> = ({ clusterId }: DataDetailsProps
       <Modal className="costManagement" isOpen={isOpen} onClose={handleOnClose} variant={ModalVariant.small}>
         <ModalHeader title={intl.formatMessage(messages.dataDetails)} />
         <ModalBody>
-          <DataDetailsContent clusterId={clusterId} />
+          <ClusterDetailsContent clusterId={clusterId} />
         </ModalBody>
       </Modal>
     </>
   );
 };
 
-export default DataDetails;
+export default ClusterDetails;
