+++ src/routes/details/components/providerDetails/clusterDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as DataDetails } from './clusterDetails';
