+++ src/routes/details/components/providerDetails/clusterDetails/utils/format.ts
