+++ src/routes/details/components/providerDetails/clusterDetails/utils/icon.tsx
