+++ src/routes/details/components/providerDetails/clusterDetails/utils/variant.ts
