+++ src/routes/details/components/providerDetails/components/sourceLink.tsx
@@ -2,7 +2,7 @@ import type { Provider } from 'api/providers';
 import messages from 'locales/messages';
 import React from 'react';
 import { useIntl } from 'react-intl';
-import { normalize } from 'routes/details/ocpBreakdown/providerDetails/utils/normailize';
+import { normalize } from 'routes/details/components/providerDetails/utils/normailize';
 import { getReleasePath } from 'utils/paths';
 
 import { styles } from './sourceLink.styles';
