+++ src/routes/details/components/providerDetails/utils/normailize.ts
