+++ src/routes/details/ocpBreakdown/ocpBreakdown.tsx
@@ -12,8 +12,8 @@ import { connect } from 'react-redux';
 import { routes } from 'routes';
 import type { BreakdownStateProps } from 'routes/details/components/breakdown';
 import { BreakdownBase } from 'routes/details/components/breakdown';
+import { DataDetails } from 'routes/details/components/providerDetails/clusterDetails';
 import { ClusterInfo } from 'routes/details/ocpBreakdown/providerDetails/clusterInfo';
-import { DataDetails } from 'routes/details/ocpBreakdown/providerDetails/dataDetails';
 import { getGroupById, getGroupByValue } from 'routes/utils/groupBy';
 import { filterProviders } from 'routes/utils/providers';
 import { getQueryState } from 'routes/utils/queryState';
