--- src/routes/details/ocpBreakdown/providerDetails/dataDetails/index.ts
@@ -1,1 +1,0 @@
-export { default as DataDetails } from './dataDetails';
